LPS-76963 Move to internal
