+++ modules/apps/foundation/portal-security-sso/portal-security-sso-openid-connect/src/main/java/com/liferay/portal/security/sso/openid/connect/internal/OpenIdConnectFlowState.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.openid.connect;
+package com.liferay.portal.security.sso.openid.connect.internal;
 
 /**
  * @author Edward C. Han
