+++ modules/apps/foundation/portal-security-sso/portal-security-sso-openid-connect/src/main/java/com/liferay/portal/security/sso/openid/connect/internal/OpenIdConnectSession.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.security.sso.openid.connect;
+package com.liferay.portal.security.sso.openid.connect.internal;
 
+import com.liferay.portal.security.sso.openid.connect.internal.OpenIdConnectFlowState;
 import com.nimbusds.oauth2.sdk.id.State;
 import com.nimbusds.oauth2.sdk.token.AccessToken;
 import com.nimbusds.oauth2.sdk.token.RefreshToken;
